🏗️ Architect: Extracted ObservationWindow from Observation class
This refactor decomposes the "God Object" Observation class by moving
complex windowing, twilight fallback, and scheduling logic into a
dedicated ObservationWindow class in apts/observations/window.py.

Key changes:
- Created ObservationWindow to handle sunset/sunrise, twilights, and time limits.
- Refactored Observation.__init__ to delegate window state to the new class.
- Removed private window-related methods from Observation class.
- Updated unit tests to correctly mock the refactored logic.
- Maintained public API consistency.
