+++ apts/observations/__init__.py
@@ -1,1 +1,4 @@
-from .base import Observation as Observation
+from .base import Observation
+from .window import ObservationWindow
+
+__all__ = ["Observation", "ObservationWindow"]
